+++ tests/common/Tag/Base/TagTest.php
@@ -71,7 +71,7 @@ final class TagTest extends TestCase
             '<test id="color" class="green">',
             TestTag::tag()
                 ->id('color')
-                ->class('red')
+                ->replaceClass('red')
                 ->attributes(['class' => 'green'])
                 ->render(),
         );
@@ -83,7 +83,7 @@ final class TagTest extends TestCase
             '<test class="green">',
             TestTag::tag()
                 ->id('color')
-                ->class('red')
+                ->replaceClass('red')
                 ->replaceAttributes(['class' => 'green'])
                 ->render(),
         );
@@ -94,7 +94,7 @@ final class TagTest extends TestCase
         $this->assertSame(
             '<test id="color" class="red">',
             TestTag::tag()
-                ->class('red')
+                ->replaceClass('red')
                 ->unionAttributes(['class' => 'green', 'id' => 'color'])
                 ->render(),
         );
@@ -151,11 +151,11 @@ final class TagTest extends TestCase
      *
      * @param string[] $class
      */
-    public function testClass(string $expected, array $class): void
+    public function testAddClass(string $expected, array $class): void
     {
         $this->assertSame($expected, (string)TestTag::tag()
-            ->class('main')
-            ->class(...$class));
+            ->addClass('main')
+            ->addClass(...$class));
     }
 
     public function dataNewClass(): array
@@ -172,7 +172,7 @@ final class TagTest extends TestCase
      */
     public function testNewClass(string $expected, ?string $class): void
     {
-        $this->assertSame($expected, (string)TestTag::tag()->class($class));
+        $this->assertSame($expected, (string)TestTag::tag()->addClass($class));
     }
 
     public function dataReplaceClass(): array
@@ -195,7 +195,7 @@ final class TagTest extends TestCase
     public function testReplaceClass(string $expected, array $class): void
     {
         $this->assertSame($expected, (string)TestTag::tag()
-            ->class('red')
+            ->replaceClass('red')
             ->replaceClass(...$class));
     }
 
@@ -208,6 +208,7 @@ final class TagTest extends TestCase
         $this->assertNotSame($tag, $tag->attribute('id', null));
         $this->assertNotSame($tag, $tag->id(null));
         $this->assertNotSame($tag, $tag->class('test'));
+        $this->assertNotSame($tag, $tag->addClass('test'));
         $this->assertNotSame($tag, $tag->replaceClass('test'));
     }
 }