+++ tests/common/Tag/H5Test.php
@@ -14,7 +14,7 @@ final class H5Test extends TestCase
         $this->assertSame(
             '<h5 class="red">Hello</h5>',
             (string) H5::tag()
-                ->class('red')
+                ->replaceClass('red')
                 ->content('Hello')
         );
     }