+++ tests/common/Tag/ITest.php
@@ -14,7 +14,7 @@ final class ITest extends TestCase
         $this->assertSame(
             '<i class="red">Hello</i>',
             (string)I::tag()
-                ->class('red')
+                ->replaceClass('red')
                 ->content('Hello')
         );
     }