+++ tests/common/Tag/LiTest.php
@@ -14,7 +14,7 @@ final class LiTest extends TestCase
         $this->assertSame(
             '<li class="red">Hello</li>',
             (string)Li::tag()
-                ->class('red')
+                ->replaceClass('red')
                 ->content('Hello')
         );
     }